refactor(tauri): split system_tray to another file
