+++ app/tauri/src/commands.rs
@@ -3,19 +3,12 @@
 
 use std::sync::Mutex;
 // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
-use tauri::{Builder, Icon, PhysicalSize, Runtime, Wry};
-use tauri::{
-    menu::{MenuBuilder, MenuItemBuilder},
-    tray::{ClickType, TrayIconBuilder},
-};
-
-use tauri::Manager;
-use base64;
-use base64::Engine;
-use base64::engine::general_purpose;
+use tauri::{Builder, PhysicalSize, Runtime, Wry};
 
 #[non_exhaustive]
 struct WindowSize;
+
+use crate::system_tray;
 
 impl WindowSize {
 
@@ -159,32 +152,9 @@ fn set_native_titlebar<R: Runtime>(use_native_titlebar: bool, window: tauri::Win
         Ok(_) => (),
         Err(e) => println!("There was a problem setting the window decorations! {:?}", e),
     }
-    window.start_dragging();
     println!("set_native_titlebar! {}", use_native_titlebar);
 }
 
-#[tauri::command]
-fn tray_icon_update<R: Runtime>(data_url: String, window: tauri::Window<R>) {
-    println!("tray_icon_update!");
-    let base64_str = data_url.trim_start_matches("data:image/png;base64,");
-
-    let decoded_vec = match general_purpose::STANDARD.decode(base64_str) {
-        Ok(vec) => vec,
-        Err(e) => {
-            eprintln!("Error decoding Base64 string: {}", e);
-            return;
-        }
-    };
-
-    let icon: Icon = Icon::Raw(decoded_vec);
-    let tray = window.app_handle().tray();
-
-    if let Some(tray) = tray {
-        if let Err(e) = tray.set_icon(Some(icon)) {
-            eprintln!("Error setting tray icon: {}", e);
-        }
-    }
-}
 
 /**
  * We could do this by passing the object into a custom function that adds the commands but I wanted
@@ -193,52 +163,14 @@ fn tray_icon_update<R: Runtime>(data_url: String, window: tauri::Window<R>) {
  * Switch to a function that takes and returns tauri::Builder<Wry> or uses a reference if we need to
  * switch it.
  */
-pub trait PomatezExtras {
+pub trait PomatezCommands {
     fn register_pomatez_commands(self) -> tauri::Builder<Wry>;
-    fn set_pomatez_system_tray(self) -> tauri::Builder<Wry>;
 }
 
-impl PomatezExtras for Builder<Wry> {
+impl PomatezCommands for Builder<Wry> {
     fn register_pomatez_commands(self) -> tauri::Builder<Wry> {
         self.invoke_handler(tauri::generate_handler![set_show, set_always_on_top,
             set_fullscreen_break, set_compact_mode, set_ui_theme, set_native_titlebar,
-            tray_icon_update, set_close, set_minimize])
-    }
-
-    fn set_pomatez_system_tray(self) -> tauri::Builder<Wry> {
-        println!("Setting system tray");
-        self.setup(|app| {
-            // Was defined in tauri.config.json to start in v1
-            // That was created with an id of 1 though this gives more control
-
-            let show = MenuItemBuilder::with_id("show", "Show").build(app);
-            let quit = MenuItemBuilder::with_id("quit", "Quit").build(app);
-            let menu = MenuBuilder::new(app).items(&[&show, &quit]).build()?;
-            let tray = TrayIconBuilder::new()
-                .menu(&menu)
-                .tooltip("Pomatez")
-                .on_menu_event(move |app, event| match event.id().as_ref() {
-                    "show" => {
-                        let window = app.get_window("main").unwrap();
-                        window.show().unwrap();
-                        window.set_focus().unwrap();
-                    }
-                    "quit" => {
-                        app.exit(0);
-                    }
-                    _ => {}
-                })
-                .on_tray_event(|tray, event| {
-                    if event.click_type == ClickType::Left {
-                        let app = tray.app_handle();
-                        let window = app.get_window("main").unwrap();
-                        window.show().unwrap();
-                        window.set_focus().unwrap();
-                    }
-                })
-                .icon(Icon::File("./icons/icon.png".into()))
-                .build(app)?;
-            Ok(())
-        })
+            system_tray::tray_icon_update, set_close, set_minimize])
     }
 }
